YAY-114 fix: transfer MeResponse, authStorage, SavaAcessToken, controlled
