+++ src/entities/post/ui/Post/PostHeader/PostMenu/PostMenu.tsx
@@ -6,7 +6,7 @@ import { Typography } from '@/shared/components/Typography'
 import { POST_MENU_ITEMS } from '@/entities/post/ui/Post/PostHeader/PostMenu/config'
 import { DeletePostModal } from '@/features/post-creator/ui/modals/DeletePostModal/DeletePostModal'
 import { usePostController } from '@/features/post/model/hooks'
-import { useMeQuery } from '@/features/auth/api'
+import { useMeQuery } from '@/entities/user/api'
 
 type Props = {
    userId: string
