+++ src/features/auth/api/auth.service.ts
@@ -1,7 +1,6 @@
 'use client'
 import { AuthEndpoints } from '@/shared/enums'
 import type {
-   MeResponse,
    SignInArgs,
    SignInResponse,
    RecoveryPasswordArgs,
@@ -15,16 +14,10 @@ import type {
 import { TOKEN } from '@/shared/constants'
 import { authChannel } from '@/shared/lib/authBroadcast'
 import { baseApi } from '@/app/store'
+import { userService } from '@/entities/user'
 
 export const authService = baseApi.injectEndpoints({
    endpoints: builder => ({
-      me: builder.query<MeResponse, void>({
-         query: () => ({
-            method: 'GET',
-            url: AuthEndpoints.me,
-         }),
-         providesTags: ['Me'],
-      }),
       signUp: builder.mutation<void, SignUpArgs>({
          query: args => ({
             method: 'POST',
@@ -72,7 +65,7 @@ export const authService = baseApi.injectEndpoints({
             const { data } = await queryFulfilled
             localStorage.setItem(TOKEN, data.accessToken)
             authChannel.postMessage({ type: 'LOGIN' })
-            dispatch(authService.endpoints.me.initiate())
+            dispatch(userService.endpoints.me.initiate())
          },
       }),
       resendEmail: builder.mutation<void, VerificationExpiredArgs>({
@@ -114,7 +107,6 @@ export const authService = baseApi.injectEndpoints({
 })
 
 export const {
-   useMeQuery,
    useSignUpMutation,
    useLogoutMutation,
    useConfirmEmailMutation,
