+++ src/features/auth/forms/SignInForm/SignInForm.tsx
@@ -4,7 +4,7 @@ import React, { useEffect } from 'react'
 import { Card } from '@/shared/components/Card'
 import { Typography } from '@/shared/components/Typography'
 import { useForm } from 'react-hook-form'
-import { ControlledInput } from '@/shared/components/Controlled'
+import { ControlledInput } from '@/shared/lib/Controlled'
 import { z } from 'zod'
 import { zodResolver } from '@hookform/resolvers/zod'
 import { Button } from '@/shared/components/Button'
