--- src/features/auth/model/index.ts
@@ -1,1 +1,0 @@
-export * from './authStorage'
