+++ src/features/post-creator/ui/modals/EditPostModal/EditPostModal.tsx
@@ -4,7 +4,7 @@ import React, { useState } from 'react'
 import { PostModal } from '@/shared/components/PostModal'
 import { useForm } from 'react-hook-form'
 import { Typography } from '@/shared/components/Typography'
-import { ControlledTextarea } from '@/shared/components/Controlled'
+import { ControlledTextarea } from '@/shared/lib/Controlled'
 import { Button } from '@/shared/components/Button'
 import { z } from 'zod'
 import { zodResolver } from '@hookform/resolvers/zod'
