+++ src/features/settings/ui/InfoTabPage/GeneralInormationForm/GeneralInormationForm.tsx
@@ -5,10 +5,10 @@ import { useForm } from 'react-hook-form'
 import { zodResolver } from '@hookform/resolvers/zod'
 import z from 'zod'
 import { generalInformationSchema } from '@/features/settings/model/schemas/GeneralInformationSchema'
-import { ControlledInput, ControlledTextarea } from '@/shared/components/Controlled'
-import { ControlledDatePicker } from '@/shared/components/Controlled/ControlledDatePicker'
+import { ControlledInput, ControlledTextarea } from '@/shared/lib/Controlled'
+import { ControlledDatePicker } from '@/shared/lib/Controlled/ControlledDatePicker'
 import { CountriesResponse, getProfileResponse } from '@/features/settings/api/settings.types'
-import { ControlledSelect } from '@/shared/components/Controlled/ControlledSelect'
+import { ControlledSelect } from '@/shared/lib/Controlled/ControlledSelect'
 import { Button } from '@/shared/components/Button'
 import { dateFormatter } from '@/shared/utils/date/dateFormatter'
