--- src/shared/components/SaveAccessToken/index.ts
@@ -1,1 +1,0 @@
-export * from './SaveAccessToken'
