+++ src/shared/lib/index.ts
@@ -1,3 +1,6 @@
 export { typographyVariants } from './typographyVariants'
 export type { TypographyVariant } from './typographyVariants'
 export { cn } from './cn'
+export { authStorage } from './authStorage'
+export { SaveAccessToken } from './SaveAccessToken'
+export * from './Controlled'
